Move resourceproxy to principal's package hierarchy
