+++ principal/resourceproxy/options.go
