+++ principal/resourceproxy/params_test.go
