+++ principal/resourceproxy/resourceproxy.go
