+++ principal/resourceproxy/resourceproxy_test.go
