+++ principal/resourceproxy/tracking.go
