+++ principal/resourceproxy/tracking_test.go
